server/requirements.txt
Django
python-dotenv
psycopg2
djangorestframework
requests
celery
redis
django-celery-beat
django-celery-results